backend_django/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg[binary]
python-dotenv
dj-database-url
PyJWT
whitenoise
pillow
gunicorn
# Serverless Dependencies
zappa
django-storages
boto3
s3transfer
